+++ packages/edit/plugin.js
@@ -30,7 +30,7 @@ K.Plugin({
 		}
 	},
 	queries: {
-		'placesEdit': function() {
+		placesEdit: function() {
 			var userId = Meteor.userId();
 			console.log('EDIT query placesEdit', userId)
 			return _.isString(userId) ? {userId: userId } : {};