+++ packages/lib/package.js
@@ -16,7 +16,7 @@ Npm.depends({
   "bootstrap-confirm-button": "0.3.2",
   "leaflet": "1.3.4",
   "leaflet-gps": "1.7.6",
-  "leaflet-layerjson": "0.3.1",
+  "leaflet-layerjson": "0.3.3",
   "leaflet.markercluster": "1.4.1",
   "leaflet-fullscreen": "1.0.2"
 });