+++ packages/stats/client/Map_stats.js
@@ -34,9 +34,11 @@ Tracker.autorun(function(comp) {
 			K.Map.layers.stats = L.geoJSON([], {
 				pointToLayer: function(point, loc) {
 					var r = point.properties.rank;
-					r = Math.min(r, 15);
-					r = Math.max(r, 5);
-					return L.circleMarker(loc, {radius: r }).on('click', function(e) {
+					var radius = Math.min(r, 15);
+						radius = Math.max(r, 5);
+					return L.circleMarker(loc, {
+							radius: radius
+					}).on('click', function(e) {
 						L.DomEvent.stopPropagation(e);
 						K.Map.map.removeLayer(K.Map.layers.stats);
 						K.Map.map.flyTo(loc, K.settings.public.map.dataMinZoom+1, {